backend/pyproject.toml
[tool.poetry]
name = "particlar1"
version = "0.1.0"
description = ""
authors = ["AlirezMortezaei <Mortezaei2324@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.27.1"
fastapi = "^0.109.2"
jose = "^1.0.0"
sqlalchemy = "^2.0.27"
sqlmodel = "^0.0.16"
celery = "^5.3.6"
redis = "^5.0.1"
pydantic = {extras = ["email"], version = "^2.6.2"}
python-jose = "^3.3.0"
passlib = "^1.7.4"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
pillow = "^10.2.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
